Add a self-pipe to signal client shutdowns
The self-pipe is "poked" when stopping a zclient, which interrupts the
select() call waiting on data. The file descriptor for the self-pipe is
stored in the 'fd' member of the 'pass_info' struct. As far as I can
tell, this member was unused, but there's no documentation on it or
tests to guarantee it.