Fix issues with self-pipe
The self-pipe used to signal a shutdown and interrupt the select() call
was being initialized too late in the code. If anything followed the
'cleanup' goto label before it was initialzed, closing the file
descriptors stored in 'self_pipe' could close stdin - which would lead
readline() to think EOF was reached, eventually quitting the
application.