Add Xcode Project
The Xcode Project is setup to build very similarly to the way the CMake
scripts build the various targets. For consistency, the same targets are
created in the Xcode Project, but only schemes to build
'reference_client' and 'reference_listener' are provided. The schemes
are setup to use Terminal.app as their console, since Xcode's Console
does not play nice with 'readline'. The schemes are also configured with
default command line arguments, which are expected to be modified
according to the development environment.

As part of the build process, 'zw_gen_hdr.c' is generated by running the
project's Python script, and the output placed in the Xcode Project's
DerivedData directory. The Z-Wave Command Class Definition XML is also
copied there, so it is readily available to the applications when
running from within Xcode.